// src/app/services/apis/emprestimo-api/interfaces/emprestimoDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface EmprestimoDTO {
    dataMovimentacao?: Date;
    id?: number;
    leitorId?: number;
    livroId?: number;
    movimentacao?: string;
}