// src/app/services/apis/leitores-api/interfaces/leitorDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface LeitorDTO {
    cpf?: string;
    id?: number;
    nome?: string;
    preferencias?: string;
}