// src/app/services/apis/livros-api/interfaces/livroDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface LivroDTO {
    ano?: string;
    autor?: string;
    editora?: string;
    id?: number;
    isbn?: string;
    titulo?: string;
}